Refactored Key Record Writers for Raw and Binary to reduce code duplication
