+++ src/java/org/apache/sqoop/mapreduce/RawKeyTextOutputFormat.java
@@ -39,48 +39,6 @@ import org.apache.hadoop.util.*;
  */
 public class RawKeyTextOutputFormat<K, V> extends FileOutputFormat<K, V> {
 
-  /**
-   * RecordWriter to write to plain text files.
-   */
-  public static class RawKeyRecordWriter<K, V> extends RecordWriter<K, V> {
-
-    private static final String UTF8 = "UTF-8";
-
-    protected DataOutputStream out;
-
-    public RawKeyRecordWriter(DataOutputStream out) {
-      this.out = out;
-    }
-
-    /**
-     * Write the object to the byte stream, handling Text as a special
-     * case.
-     * @param o the object to print
-     * @throws IOException if the write throws, we pass it on
-     */
-    private void writeObject(Object o) throws IOException {
-      if (o instanceof Text) {
-        Text to = (Text) o;
-        out.write(to.getBytes(), 0, to.getLength());
-      } else if (o instanceof BytesWritable) {
-        BytesWritable to = (BytesWritable) o;
-        out.write(to.getBytes(), 0, to.getLength());
-      } else {
-        out.write(o.toString().getBytes(UTF8));
-      }
-    }
-
-    public synchronized void write(K key, V value) throws IOException {
-      writeObject(key);
-    }
-
-    public synchronized void close(TaskAttemptContext context)
-        throws IOException {
-      out.close();
-    }
-
-  }
-
   public RecordWriter<K, V> getRecordWriter(TaskAttemptContext context)
       throws IOException {
     boolean isCompressed = getCompressOutput(context);
@@ -106,7 +64,7 @@ public class RawKeyTextOutputFormat<K, V> extends FileOutputFormat<K, V> {
       ostream = new DataOutputStream(codec.createOutputStream(fileOut));
     }
 
-    return new RawKeyRecordWriter<K, V>(ostream);
+    return new KeyRecordWriters.RawKeyRecordWriter<K, V>(ostream);
   }
 
 }
