+++ src/java/org/apache/sqoop/mapreduce/mainframe/MainframeImportJob.java
@@ -30,8 +30,8 @@ import org.apache.hadoop.mapreduce.lib.output.LazyOutputFormat;
 import org.apache.sqoop.SqoopOptions;
 import org.apache.sqoop.manager.ImportJobContext;
 
+import org.apache.sqoop.mapreduce.BinaryKeyOutputFormat;
 import org.apache.sqoop.mapreduce.DataDrivenImportJob;
-import org.apache.sqoop.mapreduce.RawKeyTextOutputFormat;
 
 /**
  * Import data from a mainframe dataset, using MainframeDatasetInputFormat.
